Fix some FindBugs stuff.

+++ src/main/java/org/drools/planner/examples/ras2012/RAS2012Solution.java
@@ -7,11 +7,10 @@ import java.util.Map;
 import java.util.SortedSet;
 import java.util.TreeSet;
 
-import org.drools.planner.examples.ras2012.model.ItineraryAssignment;
-
 import org.drools.planner.api.domain.solution.PlanningEntityCollectionProperty;
 import org.drools.planner.core.score.buildin.hardandsoft.HardAndSoftScore;
 import org.drools.planner.core.solution.Solution;
+import org.drools.planner.examples.ras2012.model.ItineraryAssignment;
 import org.drools.planner.examples.ras2012.model.MaintenanceWindow;
 import org.drools.planner.examples.ras2012.model.Network;
 import org.drools.planner.examples.ras2012.model.Route;
@@ -19,7 +18,7 @@ import org.drools.planner.examples.ras2012.model.Train;
 
 public class RAS2012Solution implements Solution<HardAndSoftScore> {
 
-    public static final Integer                   PLANNING_HORIZON_MINUTES       = 12 * 60;
+    public static final long                      PLANNING_HORIZON_MINUTES       = 12 * 60;
 
     public static final double                    PLANNING_TIME_DIVISION_MINUTES = 0.5;
